+++ src/routes/(app)/(authed)/AlbumLink.svelte
@@ -1,10 +1,11 @@
 <script lang="ts">
   import AlbumImage from '$lib/components/AlbumImage.svelte';
-  import { observeVisibility } from '$lib/observeVisibility';
+  import { observeVisibility } from '$lib/actions/observeVisibility';
   import { quintOut } from 'svelte/easing';
   import { scale } from 'svelte/transition';
   import { vibrate } from '$lib/actions/vibrate';
   import type { PageData } from './$types';
+  import { getExpressiveScheme, schemeToCSS } from '$lib/materialColors';
 
   const DELAY = 20;
 
@@ -24,26 +25,35 @@
       return DELAY * index;
     }
   });
+  let schemeStyle = $state('');
+
+  $effect(() => {
+    if (animate) {
+      getExpressiveScheme(album?.id || '', album?.albumArtId || '').then((scheme) => {
+        schemeStyle = schemeToCSS(scheme);
+      });
+    }
+  });
 </script>
 
 <div class="h-full w-full">
   {#if animate}
     <a
       use:vibrate
-      class="absolute inset-0 overflow-hidden rounded-xl"
-      class:bg-zinc-900={!album.albumArtDarkMuted}
+      class={['absolute inset-0 overflow-hidden rounded-xl', !album.albumArtId && 'bg-zinc-900']}
       href="/album/{album.id}"
       in:scale={{ duration: 400, delay, easing: quintOut, start: 0.8 }}
+      style={schemeStyle}
     >
       <AlbumImage {album} />
       <div
-        class="absolute bottom-0 left-0 flex w-full flex-col justify-end gap-0.5 p-1 text-center"
+        class="text-on-surface absolute bottom-0 left-0 flex w-full flex-col justify-end gap-0.5 p-1 text-center"
       >
         <div class="z-10 overflow-hidden rounded-[10px] whitespace-nowrap">
           <div class="absolute -top-24 -left-1 h-fit w-[calc(100%+0.5rem)] md:-top-28 xl:-top-40">
             <AlbumImage {album} maxSize="s" blur />
           </div>
-          <div class="h-full w-full overflow-hidden bg-zinc-900/80 px-2 text-ellipsis">
+          <div class="bg-surface/80 h-full w-full overflow-hidden px-2 text-ellipsis">
             {album.title}
           </div>
         </div>
@@ -51,7 +61,7 @@
           <div class="absolute -top-31 -left-1 h-fit w-[calc(100%+0.5rem)] md:-top-36 xl:-top-48">
             <AlbumImage {album} maxSize="s" blur />
           </div>
-          <div class="h-full w-full overflow-hidden bg-zinc-900/80 px-1 text-ellipsis">
+          <div class="bg-surface/80 h-full w-full overflow-hidden px-1 text-ellipsis">
             {album.albumArtist.name}
           </div>
         </div>