+++ src/routes/(app)/(authed)/profile/+page.svelte
@@ -6,14 +6,16 @@
   import RoundRefresh from '~icons/ic/round-refresh';
   import TrashFill from '~icons/iconamoon/trash-fill';
   import History from '~icons/iconamoon/history';
+  import InformationCircleFill from '~icons/iconamoon/information-circle-fill';
   import TrackRow from '$lib/components/TrackRow.svelte';
   import Accordion from '$lib/components/Accordion.svelte';
   import { getReadableTime } from '$lib/utils';
   import { flip } from 'svelte/animate';
   import type { SubmitFunction } from '@sveltejs/kit';
   import type { ActionData, PageData } from './$types';
-  import { untrack } from 'svelte';
-  import { ROLE } from '$lib/shared/consts';
+  import { onMount, untrack } from 'svelte';
+  import { ROLE, SCHEME_TYPES } from '$lib/shared/consts';
+  import Modal from '$lib/components/Modal.svelte';
 
   interface Props {
     data: PageData;
@@ -22,6 +24,8 @@
 
   let { data, form }: Props = $props();
 
+  let selectedScheme: (typeof SCHEME_TYPES)[number] = $state('EXPRESSIVE');
+  let darkMode = $state<'true' | 'false' | 'auto'>('auto');
   let deleteClicked = $state(false);
   let deleteTimeout: string | number | NodeJS.Timeout | undefined = $state();
   let loading = $state(false);
@@ -56,14 +60,27 @@
       }
     };
   };
+
+  onMount(() => {
+    selectedScheme =
+      (localStorage.getItem('schemeType') as (typeof SCHEME_TYPES)[number]) || 'EXPRESSIVE';
+    const localStorageDarkMode = localStorage.getItem('darkMode');
+    if (localStorageDarkMode !== null) {
+      darkMode = localStorageDarkMode as 'true' | 'false';
+    } else {
+      darkMode = 'auto';
+    }
+  });
 </script>
 
-<div class="absolute top-1 right-3 z-50 text-sm font-bold text-zinc-600/20">
-  App version: {data.APP_VERSION}
-</div>
 <div class="flex h-full w-full flex-col items-center overflow-auto">
   <div class="flex w-full max-w-xl flex-col p-2">
-    <div class="p-2 text-center text-xl font-bold">Profile</div>
+    <div
+      class="p-2 text-center text-xl font-bold"
+      in:fly|global={{ duration: 500, y: -10, easing: quintOut }}
+    >
+      Profile
+    </div>
 
     <Accordion title="Update Profile">
       <div class="flex w-full flex-none justify-center">
@@ -80,7 +97,7 @@
           }}
         >
           <label class="flex flex-col gap-1">
-            <div class="text-sm font-bold text-zinc-400">Username</div>
+            <div class="text-sm font-bold">Username</div>
             <input
               autocomplete="username"
               id="username"
@@ -91,7 +108,7 @@
             />
           </label>
           <label class="flex flex-col gap-1">
-            <div class="text-sm font-bold text-zinc-400">Email</div>
+            <div class="text-sm font-bold">Email</div>
             <input
               id="email"
               autocomplete="email"
@@ -103,7 +120,7 @@
             />
           </label>
           <button
-            class="bg-primary text-accessible hover:bg-primary/80 disabled:bg-primary mt-2 w-full self-center rounded-md px-4 py-1 font-semibold transition-colors disabled:opacity-50"
+            class="bg-primary text-on-primary hover:bg-primary/80 disabled:bg-primary mt-2 w-full self-center rounded-md px-4 py-1 font-semibold transition-colors disabled:opacity-50"
             type="submit"
             use:vibrate
             disabled={loading}
@@ -138,7 +155,7 @@
           }}
         >
           <label class="flex flex-col gap-1">
-            <div class="text-sm font-bold text-zinc-400">Current password</div>
+            <div class="text-sm font-bold">Current password</div>
             <input
               autocomplete="current-password"
               type="password"
@@ -149,7 +166,7 @@
             />
           </label>
           <label class="flex flex-col gap-1">
-            <div class="text-sm font-bold text-zinc-400">New password</div>
+            <div class="text-sm font-bold">New password</div>
             <input
               id="newpassword"
               autocomplete="new-password"
@@ -160,7 +177,7 @@
             />
           </label>
           <label class="flex flex-col gap-1">
-            <div class="text-sm font-bold text-zinc-400">Repeat new password</div>
+            <div class="text-sm font-bold">Repeat new password</div>
             <input
               id="repeatpassword"
               autocomplete="new-password"
@@ -174,7 +191,7 @@
             <div class="text-sm font-bold text-red-500">{form.error}</div>
           {/if}
           <button
-            class="bg-primary text-accessible hover:bg-primary/80 disabled:bg-primary mt-2 w-full self-center rounded-md px-4 py-1 font-semibold transition-colors disabled:opacity-50"
+            class="bg-primary text-on-primary hover:bg-primary/80 disabled:bg-primary mt-2 w-full self-center rounded-md px-4 py-1 font-semibold transition-colors disabled:opacity-50"
             type="submit"
             use:vibrate
             disabled={loading}
@@ -191,53 +208,98 @@
       </div>
     </Accordion>
 
+    <Accordion title="Color Scheme" delay={200}>
+      <div class="flex w-full flex-col gap-2 p-4">
+        <div class="text-sm font-bold">Choose a color scheme for the app</div>
+        <select
+          class="focus-visible:ring-primary focus-within:ring-primary w-full rounded-xl border-none bg-zinc-600/20 py-1 outline-hidden transition-all hover:bg-zinc-600/50 focus-visible:ring-2"
+          onchange={(e) => {
+            const value = (e.target as HTMLSelectElement).value;
+            localStorage.setItem('schemeType', value);
+            location.reload();
+          }}
+          bind:value={selectedScheme}
+        >
+          {#each SCHEME_TYPES as schemeType}
+            <option value={schemeType}>
+              {schemeType.charAt(0) + schemeType.slice(1).toLowerCase().replaceAll('_', ' ')}
+              {#if schemeType === 'EXPRESSIVE'}
+                (default)
+              {/if}
+            </option>
+          {/each}
+        </select>
+
+        <div class="text-sm font-bold">Dark mode</div>
+        <select
+          class="focus-visible:ring-primary focus-within:ring-primary w-full rounded-xl border-none bg-zinc-600/20 py-1 outline-hidden transition-all hover:bg-zinc-600/50 focus-visible:ring-2"
+          onchange={(e) => {
+            const value = (e.target as HTMLSelectElement).value;
+            if (value === 'auto') {
+              localStorage.removeItem('darkMode');
+            } else {
+              localStorage.setItem('darkMode', value);
+            }
+            location.reload();
+          }}
+          value={darkMode}
+        >
+          <option value="auto">Auto (follow system preference)</option>
+          <option value="true">Dark</option>
+          <option value="false">Light</option>
+        </select>
+      </div>
+    </Accordion>
+
     {#if data.user?.role !== ROLE.OWNER}
       <div
+        in:fly|global={{ duration: 500, x: -20, easing: quintOut, delay: 300 }}
         class="mt-4 flex w-full items-center justify-between rounded-xl bg-zinc-600/20 py-2 pr-2 pl-4"
       >
-        <div
-          in:fly|global={{ duration: 500, x: -20, easing: quintOut, delay: 500 }}
-          class="text-center text-xl font-bold"
-        >
-          Delete account
-        </div>
+        <div class="text-center text-xl font-bold">Delete account</div>
 
-        <div
-          in:fly|global={{ duration: 500, x: -20, easing: quintOut, delay: 600 }}
-          class="flex items-center justify-center"
-        >
-          <form
-            method="POST"
-            use:enhance={({ cancel }) => {
-              if (!deleteClicked) {
-                deleteClicked = true;
-                deleteTimeout = setTimeout(() => {
-                  deleteClicked = false;
-                }, 2000);
-                cancel();
-              } else {
-                deleteClicked = false;
-                clearTimeout(deleteTimeout);
-              }
-            }}
-            class="flex w-full items-center justify-center"
-            action="?/delete"
+        <div class="flex items-center justify-center">
+          <button
+            onclick={() => (deleteClicked = true)}
+            class="w-full rounded-md bg-rose-600 px-4 py-1 font-semibold text-white transition-all"
+            use:vibrate
           >
-            <button
-              type="submit"
-              class="hover:bg-opacity-80 w-full rounded-md bg-rose-700 px-4 py-1 font-semibold transition-all"
-            >
-              {#if deleteClicked}
-                Are you sure?
-              {:else}
-                <TrashFill class="text-xl" />
-              {/if}
-            </button>
-          </form>
+            <TrashFill class="text-2xl" />
+          </button>
         </div>
       </div>
     {/if}
   </div>
+
+  {#if deleteClicked}
+    <Modal title="Are you sure?" onclose={() => (deleteClicked = false)}>
+      <div class="flex h-full flex-col items-center justify-center gap-10 p-6">
+        <div class="rounded-xl bg-rose-600/10 p-4 text-center text-sm font-bold text-rose-600">
+          <InformationCircleFill class="inline align-top text-base" />
+          This action cannot be undone and will delete all your data, including your listening history!
+        </div>
+
+        <div class="flex w-full items-center justify-center gap-4 max-sm:flex-col">
+          <form method="POST" action="?/delete">
+            <button
+              type="submit"
+              class="rounded-3xl bg-rose-600 px-4 py-2 font-semibold text-white transition-all"
+              use:vibrate
+            >
+              Yes, delete my account
+            </button>
+          </form>
+          <button
+            class="bg-primary text-on-primary rounded-3xl px-4 py-2 font-semibold transition-all"
+            onclick={() => (deleteClicked = false)}
+            use:vibrate
+          >
+            No, keep my account
+          </button>
+        </div>
+      </div>
+    </Modal>
+  {/if}
 
   {#if data.listens.length > 0}
     <div